Create EmployeeEvaluationHelper and move the EmployeeEvaluationDTO mapping to it
