+++ source/CommonJobs/CommonJobs.Application.EvalForm/CommonJobs.Application.Evaluations.csproj
@@ -78,6 +78,7 @@
     <Compile Include="Dtos\UpdateCalificationDto.cs" />
     <Compile Include="Dtos\UpdateEvaluationDto.cs" />
     <Compile Include="Commands\RevertEvaluationStateCommand.cs" />
+    <Compile Include="Helper\EmployeeEvaluationHelper.cs" />
     <Compile Include="Helper\EvaluationStateHelper.cs" />
     <Compile Include="Helper\RevertEvaluationActionsHelper.cs" />
     <Compile Include="Indexes\EmployeeToEvaluate_Search.cs" />
